Did something

+++ src/ALU.java
@@ -8,34 +8,34 @@ public class ALU {
         flags = new Flags();
     }
 
-    public byte add(byte first, byte second) {
+    public short add(short first, short second) {
         int f,s,r;
         f = first; s = second; r = first+second;
         return flags.adjust(r);
     }
 
-    public byte addWithCarry(byte first, byte second) {
+    public short addWithCarry(short first, short second) {
         int r = (int)add(first, second);
         if (flags.getFlag("carry"))
             r++;
         return flags.adjust(r);
     }
 
-    public byte twoscomplement(byte b) {
+    public short twoscomplement(short b) {
         int bt = b;
         bt = (0xFF-bt)+1;
         return flags.adjust(bt);
     }
 
-    public byte subtract(byte first, byte second) {
+    public short subtract(short first, short second) {
         return add(first,twoscomplement(second));
     }
 
-    public byte subtractWithBorrow(byte first, byte second) {
+    public short subtractWithBorrow(short first, short second) {
         boolean bor = flags.getFlag("carry");
-        byte r = subtract(first,second);
+        short r = subtract(first,second);
         if (bor)
-            return subtract(r,(byte)0x01);
+            return subtract(r,(short)0x01);
         return r;
     }
 }
